refactor(web): separate Resume browser state
