+++ examples/web/src/features/resume/browser/use-conversation-selection-scroll.ts
@@ -1,0 +1,47 @@
+import { useEffect, useRef } from 'react';
+
+export function useConversationSelectionScroll(
+  isConversationVisible: boolean,
+  selectedEntryId: string,
+) {
+  const conversationPanelRef = useRef<HTMLElement | null>(null);
+  const selectedConversationRef = useRef<HTMLLIElement | null>(null);
+  const revealConversationSelectionRef = useRef(false);
+
+  useEffect(() => {
+    if (!isConversationVisible) return;
+    const frame = window.requestAnimationFrame(() => {
+      const panel = conversationPanelRef.current;
+      const target = selectedConversationRef.current;
+      if (panel === null || target === null) return;
+      const usesInternalScroll = window.getComputedStyle(panel).overflowY !== 'visible';
+      if (usesInternalScroll) {
+        const panelBox = panel.getBoundingClientRect();
+        const targetBox = target.getBoundingClientRect();
+        const headingHeight = panel.querySelector<HTMLElement>('.pilot-panel-heading')
+          ?.getBoundingClientRect().height ?? 0;
+        panel.scrollTo({
+          top: Math.max(
+            0,
+            panel.scrollTop + targetBox.top - panelBox.top - headingHeight - 4,
+          ),
+          behavior: 'auto',
+        });
+      } else if (revealConversationSelectionRef.current) {
+        target.scrollIntoView({ block: 'center', behavior: 'auto' });
+      }
+      revealConversationSelectionRef.current = false;
+    });
+    return () => window.cancelAnimationFrame(frame);
+  }, [isConversationVisible, selectedEntryId]);
+
+  const revealConversationSelection = (): void => {
+    revealConversationSelectionRef.current = true;
+  };
+
+  return {
+    conversationPanelRef,
+    revealConversationSelection,
+    selectedConversationRef,
+  };
+}
